requirements.txt
django
djangorestframework
psycopg2
gunicorn
whitenoise
djoser
django-cors-headers
django-storages[s3]
gspread
oauth2client
drf-spectacular